refactor: 타입스크립트 추가 적용
