+++ package.json
@@ -7,9 +7,10 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.0.1",
     "@testing-library/user-event": "^14.1.1",
+    "@types/react": "^18.2.15",
     "axios": "^1.4.0",
     "bootstrap": "4.6.2",
-    "formik": "^2.4.1",
+    "formik": "^2.4.2",
     "moment": "^2.29.4",
     "query-string": "^8.1.0",
     "react": "^18.2.0",
