+++ src/app/data/consultingDetailData.ts
