+++ src/app/data/lectureData.ts
