+++ src/app/data/lectureDetailData.ts
