+++ src/app/pages/FindPassword/FindPassword.tsx
@@ -6,10 +6,16 @@ import * as Yup from "yup";
 import { clearMessage } from "app/slices/message";
 
 import FormField from "app/components/FormField/FormField";
-import "app/pages/FindId/FindId.css";
+import "app/pages/FindPassword/FindPassword.css";
 
-const FindId = () => {
-  const { message } = useSelector((state) => state.message);
+interface MessageType {
+  message: {
+    message: string;
+  };
+}
+
+const FindPassword = () => {
+  const { message } = useSelector((state: MessageType) => state.message);
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -18,7 +24,7 @@ const FindId = () => {
 
   const initialValues = {
     phoneNumber: "",
-    password: "",
+    email: "",
   };
 
   const validationSchema = Yup.object().shape({
@@ -28,37 +34,19 @@ const FindId = () => {
         /^(010)[0-9]{3,4}[0-9]{4}$/,
         "유효하지 않은 휴대폰 번호 형식입니다."
       ),
-    password: Yup.string()
-      .required("비밀번호를 입력해주세요.")
-      .min(8, "비밀번호는 최소 8자리 이상 입력해주세요.")
-      .test(
-        "password",
-        "비밀번호는 문자, 숫자, 특수문자를 모두 포함해야 합니다.",
-        (value) => {
-          const optionalSpeciesCount = [
-            /[A-Z]/.test(value),
-            /[a-z]/.test(value),
-            /[\u3131-\uD79D]/.test(value),
-          ].filter(Boolean)?.length;
-
-          const essentialSpeciesCount = [
-            /[@$!%*?&]/.test(value),
-            /\d/.test(value),
-          ].filter(Boolean)?.length;
-
-          return optionalSpeciesCount >= 1 && essentialSpeciesCount === 2;
-        }
-      ),
+    email: Yup.string()
+      .email("이메일 형식에 맞지 않습니다.")
+      .required("이메일을 입력해주세요."),
   });
 
   return (
     <div>
       <Formik initialValues={initialValues} validationSchema={validationSchema}>
-        {({ values, errors, touched }) => (
-          <Form values={values}>
-            <div className="find-id-container">
+        {({ errors, touched }) => (
+          <Form>
+            <div className="find-pw-container">
               <div className="col-md-6 offset-md-3 col-xs-12">
-                <h2 className="text-xs-left">휴대폰 번호로 아이디 찾기</h2>
+                <h2 className="text-xs-left">휴대폰 번호로 비밀번호 찾기</h2>
                 <FormField
                   label="휴대폰 번호"
                   placeholder="휴대폰 번호를 입력해주세요."
@@ -67,17 +55,34 @@ const FindId = () => {
                   errors={errors}
                   touched={touched}
                 />
+                <div className="form-group">
+                  <button className="btn btn-lg btn-primary pull-xs-right">
+                    <span>비밀번호 찾기</span>
+                  </button>
+                </div>
+              </div>
+            </div>
+          </Form>
+        )}
+      </Formik>
+
+      <Formik initialValues={initialValues} validationSchema={validationSchema}>
+        {({ errors, touched }) => (
+          <Form>
+            <div className="find-pw-container second">
+              <div className="col-md-6 offset-md-3 col-xs-12">
+                <h2 className="text-xs-left">이메일로 비밀번호 찾기</h2>
                 <FormField
-                  label="비밀번호"
-                  placeholder="문자, 숫자, 특수문자를 모두 포함시켜주세요."
-                  name="password"
-                  type="password"
+                  label="이메일"
+                  placeholder="이메일을 입력해주세요."
+                  name="email"
+                  type="text"
                   errors={errors}
                   touched={touched}
                 />
                 <div className="form-group">
                   <button className="btn btn-lg btn-primary pull-xs-right">
-                    <span>아이디 찾기</span>
+                    <span>비밀번호 찾기</span>
                   </button>
                 </div>
               </div>
@@ -97,4 +102,4 @@ const FindId = () => {
   );
 };
 
-export default FindId;
+export default FindPassword;
